backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
litellm==1.17.9
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
pyyaml==6.0.1
hypothesis==6.92.1
pytest==7.4.3
pytest-asyncio==0.21.1